Since this version DB is creating without default values. If network response is fail and DB is empty default values from assets are send to DB. Bug with flickering of previous singleHero after clicking on singleHero without description was fixed (note: it can be fixed more optimally with flow between repository and ViewModel). HeroMoshi and other network data classes were renamed to HeroDTO and similar and moved to models package
