+++ app/src/main/java/com/example/marvel_app_project/data/HeroDao.kt
@@ -4,7 +4,7 @@ import androidx.room.Dao
 import androidx.room.Query
 import androidx.room.Update
 import androidx.room.Upsert
-import kotlinx.coroutines.flow.Flow
+import com.example.marvel_app_project.models.data.HeroEntity
 
 @Dao
 interface HeroDao {
