+++ app/src/main/java/com/example/marvel_app_project/domain/ChooseHeroesDomainState.kt
@@ -1,6 +1,6 @@
 package com.example.marvel_app_project.domain
 
-import com.example.marvel_app_project.data.HeroEntity
+import com.example.marvel_app_project.models.data.HeroEntity
 
 sealed interface ChooseHeroesDomainState {
     data class Success(val heroValues: List<HeroEntity>):ChooseHeroesDomainState
