--- app/src/main/java/com/example/marvel_app_project/models/DataLayer/MoshiResponse.kt
@@ -1,5 +1,0 @@
-package com.example.marvel_app_project.models.DataLayer
-
-data class MoshiResponse(
-    val data: MoshiResponseData
-)
