+++ app/src/main/java/com/example/marvel_app_project/models/data/network/HeroDTOResponse.kt
@@ -1,0 +1,5 @@
+package com.example.marvel_app_project.models.data.network
+
+data class HeroDTOResponse(
+    val data: HeroDTOResponseData
+)
