+++ app/src/main/java/com/example/marvel_app_project/models/data/network/HeroDTOResponseData.kt
@@ -1,8 +1,8 @@
-package com.example.marvel_app_project.models.DataLayer
+package com.example.marvel_app_project.models.data.network
 
 import com.squareup.moshi.Json
 
-data class MoshiResponseData(
+data class HeroDTOResponseData(
     @Json(name = "results")
     val result: List<HeroMoshi>
 )
