+++ app/src/main/java/com/example/marvel_app_project/network/either/Either.kt
@@ -1,4 +1,4 @@
-package com.example.marvel_app_project.network.Either
+package com.example.marvel_app_project.network.either
 
 sealed class Either<out F, out S> {
