+++ app/src/main/java/com/example/marvel_app_project/ui/pages/singlehero/SingleHeroUiState.kt
@@ -1,6 +1,6 @@
 package com.example.marvel_app_project.ui.pages.singlehero
 
-import com.example.marvel_app_project.models.UiLayer.HeroUI
+import com.example.marvel_app_project.models.ui.HeroUI
 
 sealed interface SingleHeroUiState{
     data class Success(val singleHeroUIValue: HeroUI): SingleHeroUiState
